+++ CertificateSDK.xcframework/ios-arm64_i386_x86_64-simulator/CertificateSDK.framework/Headers/CertificateRequestProtocol.h
@@ -1,0 +1,63 @@
+/*
+ CertificateRequestProtocol.h
+
+ https://github.com/jamf/CertificateSDK
+
+ MIT License
+
+ Copyright (c) 2019 Jamf Open Source Community
+
+ Permission is hereby granted, free of charge, to any person obtaining a copy
+ of this software and associated documentation files (the "Software"), to deal
+ in the Software without restriction, including without limitation the rights
+ to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ copies of the Software, and to permit persons to whom the Software is
+ furnished to do so, subject to the following conditions:
+
+ The above copyright notice and this permission notice shall be included in all
+ copies or substantial portions of the Software.
+
+ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ SOFTWARE.
+*/
+
+#ifndef CertificateRequestProtocol_h
+#define CertificateRequestProtocol_h
+
+#import <Foundation/Foundation.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+/**
+ This is the protocol used for requesting and renewing certificates from Jamf Pro.
+ */
+@protocol CertificateRequestProtocol
+
+/**
+ This is the maximum number of steps in the progress of a certificate request.
+ May be approximate until the actual call of -startNewCertificateRequest
+ */
+@property (assign, readonly) NSUInteger maxNumberOfSteps;
+
+/**
+ This will start a request to the Jamf Pro server for a new certificate.
+ The delegate will be called with progress, errors, and the new identity.
+ NOTE: Only one such request should be in progress at a given time for a single CertificateRequest object.
+ */
+- (void)startNewCertificateRequest;
+
+/**
+ Cancels the current request, if any.  The completion handler WILL be called on the delegate.
+ */
+- (void)cancelRequest;
+
+@end
+
+NS_ASSUME_NONNULL_END
+
+#endif /* CertificateRequestProtocol_h */